refactor: extract formatRelativeIgnoreFile
