+++ src/generateIgnoreFile.js
@@ -4,10 +4,11 @@ const path = require('path')
 const R = require('ramda')
 
 const decodeIgnoreSyncFile = require('./decodeIgnoreSyncFile')
+const formatRelativeIgnoreFile = require('./utils/formatRelativeIgnoreFile')
 const github = require('./utils/github')
 const highlightComments = require('./utils/highlightComments')
 const joinLinesWithEOF = require('./utils/joinLinesWithEOF')
-const { COMMENT_HEADER_ALERT, LINE_BREAK } = require('./constants')
+const { COMMENT_HEADER_ALERT } = require('./constants')
 const { dynamicComposeP, promiseMap } = require('./utils/ramdaHelper')
 const { readFile } = require('./utils/fsHelper')
 
@@ -34,26 +35,8 @@ const localSourceFetcher = async (block, directory) => {
 }
 const relativeSourceFetcher = async (block, directory) => {
   const files = await promiseMap(async (relativeFilePath) => {
-    const filePath = path.dirname(relativeFilePath)
     const fileContent = await readFile(path.join(directory, relativeFilePath))
-    const splittedFileContent = fileContent.split(LINE_BREAK)
-    const edittedFileContent = splittedFileContent.map((line) => {
-      if (!line) {
-        return line
-      }
-
-      if (line.startsWith('#')) {
-        return line
-      }
-
-      if (line.startsWith('!')) {
-        return '!' + path.join(filePath, line.substring(1))
-      }
-
-      return path.join(filePath, line)
-    })
-
-    return edittedFileContent.join(LINE_BREAK)
+    return formatRelativeIgnoreFile(fileContent, path.dirname(relativeFilePath))
   }, block.data)
   return joinLinesWithEOF(files)
 }
